requirements.txt
# ── Core API ──────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-dotenv>=1.0.0

# ── LLM / Classifier ─────────────────────────────────────────────
httpx>=0.27.0
sentence-transformers>=2.7.0
tiktoken>=0.7.0
numpy>=1.26.0

# ── Budget Guard ─────────────────────────────────────────────────
aiosqlite>=0.20.0

# ── Memory ───────────────────────────────────────────────────────
chromadb>=0.5.0

# ── Tools ────────────────────────────────────────────────────────
duckduckgo-search>=6.0.0
pymupdf>=1.24.0
pandas>=2.2.0
tabulate>=0.9.0

# ── Tests ────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0